added some helpers
